'nezha'

+++ ScholarNet/routes/web.php
@@ -19,15 +19,14 @@ use App\Models\Publication;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
-
 Route::get('/',[homeController::class,'index'])->name('home');
-Route::get('publications/addPublicaion',[publicationController::class,'create'])->name('publication.create');
-Route::post('publications/storePublication',[publicationController::class,'store'])->name('publication.store');
-Route::delete('publications/delete/{publication}',[publicationController::class,'destroy'])->name('publication.destroy');
+Route::get('publications/addPublicaion',[publicationController::class,'create'])->name('publication.create')->middleware('auth');
+Route::post('publications/storePublication',[publicationController::class,'store'])->name('publication.store')->middleware('auth');
+Route::delete('publications/delete/{publication}',[publicationController::class,'destroy'])->name('publication.destroy')->middleware('auth');
 
 
 Route::get('/login', [UserController::class, 'showLogin'])->name('showlogin');
 Route::post('/login', [UserController::class, 'login'])->name('login');
 
-Route::get('/adduser', [UserController::class, 'showadduser'])->name('adduser');
-Route::post('/adduser', [UserController::class, 'adduser'])->name('adduser');
+Route::get('/adduser', [UserController::class, 'showadduser'])->name('adduser')->middleware('auth');
+Route::post('/adduser', [UserController::class, 'adduser'])->name('adduser')->middleware('auth');
